create new emonpy_core module with EmonPyCore class containing methods removed from EmonApiCore
